--- a/benchmark/src/main/scala/fs2/benchmark/ChunkBenchmark.scala
+++ b/benchmark/src/main/scala/fs2/benchmark/ChunkBenchmark.scala
@@ -35,7 +35,7 @@ class ChunkBenchmark {
     chunkSeq = Seq.range(0, chunkCount).map(_ => Chunk.seq(Seq.fill(chunkSize)(Obj.create)))
     sizeHint = chunkSeq.foldLeft(0)(_ + _.size)
   }
-    
+
   @Benchmark
   def concat(): Unit =
     Chunk.concat(chunkSeq, sizeHint)