I ran `jmh:run -i 3 -wi 3 -f1 -t1 fs2.benchmark.ChunkBenchmark` for benchmarking.

I tried three options:
- before (current implementation)
- after (concatting done via `Array[Any]`)
- sizeHint (concatting done via `Array[Any]` + a `totalSize` parameter that avoids doing the `foldLeft` to determine the total size)

Here are the results:
```
before
[info] Benchmark              (chunkCount)  (chunkSize)   Mode  Cnt        Score        Error  Units
[info] ChunkBenchmark.concat             1           16  thrpt    3  3284142.155 ± 126490.513  ops/s
[info] ChunkBenchmark.concat             1          256  thrpt    3   473260.526 ±   7160.038  ops/s
[info] ChunkBenchmark.concat             1         4096  thrpt    3    32789.284 ±   1017.692  ops/s
[info] ChunkBenchmark.concat            20           16  thrpt    3   340847.264 ±   9499.195  ops/s
[info] ChunkBenchmark.concat            20          256  thrpt    3    24935.817 ±   1613.729  ops/s
[info] ChunkBenchmark.concat            20         4096  thrpt    3     1220.534 ±    493.120  ops/s
[info] ChunkBenchmark.concat            50           16  thrpt    3   126833.000 ±   3212.293  ops/s
[info] ChunkBenchmark.concat            50          256  thrpt    3     8982.511 ±    358.815  ops/s
[info] ChunkBenchmark.concat            50         4096  thrpt    3      430.554 ±     10.007  ops/s
[info] ChunkBenchmark.concat           100           16  thrpt    3    73317.013 ±   2234.364  ops/s
[info] ChunkBenchmark.concat           100          256  thrpt    3     4441.279 ±    413.893  ops/s
[info] ChunkBenchmark.concat           100         4096  thrpt    3      214.331 ±      4.257  ops/s
[info] ChunkBenchmark.concat          2000            1  thrpt    3    44618.582 ±    137.189  ops/s

after
[info] Benchmark              (chunkCount)  (chunkSize)   Mode  Cnt        Score        Error  Units
[info] ChunkBenchmark.concat             1           16  thrpt    3  3756793.938 ±  96298.642  ops/s
[info] ChunkBenchmark.concat             1          256  thrpt    3   726011.970 ±  38789.971  ops/s
[info] ChunkBenchmark.concat             1         4096  thrpt    3    50393.746 ±   1812.333  ops/s
[info] ChunkBenchmark.concat            20           16  thrpt    3   604061.751 ± 158657.113  ops/s
[info] ChunkBenchmark.concat            20          256  thrpt    3    44660.055 ±  17961.989  ops/s
[info] ChunkBenchmark.concat            20         4096  thrpt    3     2965.940 ±    699.276  ops/s
[info] ChunkBenchmark.concat            50           16  thrpt    3   320287.385 ±  13561.885  ops/s
[info] ChunkBenchmark.concat            50          256  thrpt    3    23875.315 ±   1041.319  ops/s
[info] ChunkBenchmark.concat            50         4096  thrpt    3      713.254 ±     18.094  ops/s
[info] ChunkBenchmark.concat           100           16  thrpt    3   163875.550 ±   3890.070  ops/s
[info] ChunkBenchmark.concat           100          256  thrpt    3    12110.938 ±    500.191  ops/s
[info] ChunkBenchmark.concat           100         4096  thrpt    3      349.634 ±      3.853  ops/s
[info] ChunkBenchmark.concat          2000            1  thrpt    3    62588.931 ±    357.599  ops/s

sizeHint
[info] Benchmark              (chunkCount)  (chunkSize)   Mode  Cnt        Score        Error  Units
[info] ChunkBenchmark.concat             1           16  thrpt    3  4109142.224 ± 232218.964  ops/s
[info] ChunkBenchmark.concat             1          256  thrpt    3   998124.583 ±  43709.896  ops/s
[info] ChunkBenchmark.concat             1         4096  thrpt    3    71019.438 ±   4468.760  ops/s
[info] ChunkBenchmark.concat            20           16  thrpt    3   602900.476 ±  29613.950  ops/s
[info] ChunkBenchmark.concat            20          256  thrpt    3    41592.014 ±    233.699  ops/s
[info] ChunkBenchmark.concat            20         4096  thrpt    3     2713.213 ±    354.821  ops/s
[info] ChunkBenchmark.concat            50           16  thrpt    3   339112.519 ±  18307.706  ops/s
[info] ChunkBenchmark.concat            50          256  thrpt    3    19158.766 ±     83.844  ops/s
[info] ChunkBenchmark.concat            50         4096  thrpt    3      573.971 ±     27.628  ops/s
[info] ChunkBenchmark.concat           100           16  thrpt    3   174644.854 ±   8583.491  ops/s
[info] ChunkBenchmark.concat           100          256  thrpt    3    11933.488 ±    157.509  ops/s
[info] ChunkBenchmark.concat           100         4096  thrpt    3      351.364 ±      3.730  ops/s
[info] ChunkBenchmark.concat          2000            1  thrpt    3    99242.645 ±   9539.513  ops/s
```

As @ChristopherDavenport  pointed out, caveat to this PR is that it's unclear how `Array[Any]` looks like in memory. If it just becomes `Array[Object]` then all's good (as you can see `A` already been determined to be a boxed object).