@mpilquist 2019-12-14
Looks good -- need to run `fmt` and correct the couple scaladoc comments and then ready for merge.